# backend/docs/swagger.yaml
components:
  schemas:
    Aggregation:
      description: Aggregation represents an aggregation result
      properties:
        count:
          description: Count in the bucket
          type: integer
        key:
          description: Bucket key
          type: string
        label:
          description: Label corresponding to the key
          type: string
      required:
      - count
      - key
      type: object
    AggregationBodyWithSqon:
      properties:
        field:
          type: string
        size:
          type: integer
        sqon:
          $ref: '#/components/schemas/Sqon'
      type: object
    ApiError:
      description: ApiError represents an error
      properties:
        detail: {}
        message:
          type: string
        status:
          type: integer
      required:
      - message
      - status
      type: object
...
    Count:
      description: Count represents count result
      properties:
        count:
          description: Number of results
          type: integer
      type: object
    CountBodyWithSqon:
      properties:
        sqon:
          $ref: '#/components/schemas/Sqon'
      type: object
...
    ExpandedSomaticSNVOccurrence:
      properties:
        aa_change:
          type: string
        ad_alt:
          type: integer
        ad_ratio:
          type: number
        ad_total:
          type: integer
        cadd_phred:
          type: number
        cadd_score:
          type: number
        chromosome:
          type: string
        clinvar:
          items:
            type: string
          type: array
          uniqueItems: false
        dann_score:
          type: number
        dna_change:
          type: string
        end:
          type: integer
        ensembl_gene_id:
          type: string
        exon_rank:
          type: integer
        exon_total:
          type: integer
        fathmm_pred:
          type: string
        fathmm_score:
          type: number
        filter:
          type: string
        gnomad_loeuf:
          type: number
        gnomad_pli:
          type: number
        gnomad_v3_af:
          type: number
        hgvsg:
          type: string
        interpretation_classification_counts:
          $ref: '#/components/schemas/JsonMap-string-int'
        is_canonical:
          type: boolean
        is_mane_plus:
          type: boolean
        is_mane_select:
          type: boolean
        locus:
          type: string
        locus_id:
          type: string
        lrt_pred:
          type: string
        lrt_score:
          type: number
        omim_conditions:
          items:
            $ref: '#/components/schemas/OmimGenePanel'
          type: array
          uniqueItems: false
        picked_consequences:
          items:
            type: string
          type: array
          uniqueItems: false
        polyphen2_hvar_pred:
          type: string
        polyphen2_hvar_score:
          type: number
        qd:
          type: number
        revel_score:
          type: number
        rsnumber:
          type: string
        sift_pred:
          type: string
        sift_score:
          type: number
        somatic_pc_tn_wgs:
          type: integer
        somatic_pf_tn_wgs:
          type: number
        somatic_pn_tn_wgs:
          type: integer
        spliceai_ds:
          type: number
        spliceai_type:
          items:
            type: string
          type: array
          uniqueItems: false
        start:
          type: integer
        symbol:
          type: string
        transcript_id:
          type: string
        vep_impact:
          $ref: '#/components/schemas/VepImpact'
      required:
      - chromosome
      - end
      - hgvsg
      - locus
      - locus_id
      - start
      type: object
...
    JsonMap-string-int:
      additionalProperties:
        type: integer
      type: object
    LeafContent:
      properties:
        field:
          type: string
        value:
          items: {}
          type: array
          uniqueItems: false
      type: object
...
    ListBodyWithSqon:
      description: Body of a list request
      properties:
        additional_fields:
          items:
            type: string
          type: array
          uniqueItems: false
        limit:
          type: integer
        offset:
          type: integer
        page_index:
          type: integer
        sort:
          items:
            $ref: '#/components/schemas/SortBody'
          type: array
          uniqueItems: false
        sqon:
          $ref: '#/components/schemas/Sqon'
      type: object
...
    OmimGenePanel:
      properties:
        inheritance_code:
          items:
            type: string
          type: array
          uniqueItems: false
        omim_phenotype_id:
          type: string
        panel:
          type: string
      type: object
...
    SomaticSNVOccurrence:
      properties:
        aa_change:
          type: string
        ad_ratio:
          type: number
        clinvar:
          items:
            type: string
          type: array
          uniqueItems: false
        germline_pc_wgs:
          type: integer
        germline_pf_wgs:
          type: number
        gnomad_v3_af:
          type: number
        has_interpretation:
          type: boolean
        has_note:
          type: boolean
        hgvsg:
          type: string
        hotspot:
          type: boolean
        is_canonical:
          type: boolean
        is_mane_plus:
          type: boolean
        is_mane_select:
          type: boolean
        locus_id:
          type: string
        omim_inheritance_code:
          items:
            type: string
          type: array
          uniqueItems: false
        picked_consequences:
          items:
            type: string
          type: array
          uniqueItems: false
        rsnumber:
          type: string
        seq_id:
          type: integer
        somatic_pc_tn_wgs:
          type: integer
        somatic_pf_tn_wgs:
          type: number
        symbol:
          type: string
        task_id:
          type: integer
        transcript_id:
          type: string
        variant_class:
          type: string
        vep_impact:
          $ref: '#/components/schemas/VepImpact'
      required:
      - aa_change
      - clinvar
      - germline_pc_wgs
      - germline_pf_wgs
      - gnomad_v3_af
      - has_interpretation
      - has_note
      - hgvsg
      - hotspot
      - is_canonical
      - is_mane_plus
      - is_mane_select
      - locus_id
      - omim_inheritance_code
      - picked_consequences
      - rsnumber
      - seq_id
      - somatic_pc_tn_wgs
      - somatic_pf_tn_wgs
      - symbol
      - task_id
      - variant_class
      - vep_impact
      type: object
    SortBody:
      properties:
        field:
          type: string
        order:
          enum:
          - asc
          - desc
          type: string
      type: object
    Sqon:
      properties:
        content:
          oneOf:
          - $ref: '#/components/schemas/LeafContent'
          - items:
              $ref: '#/components/schemas/Sqon'
            type: array
        id:
          type: string
        op:
          enum:
          - in
          - and
          - or
          - not
          - between
          - '>'
          - <
          - '>='
          - <=
          - not-in
          - all
          type: string
      type: object
    Statistics:
      description: Statistics represents statistics about a column
      properties:
        max:
          description: Max for numeric facet
          type: number
        min:
          description: Min for numeric facet
          type: number
        type:
          description: To tell UI if it's an integer or a decimal
          type: string
      type: object
    StatisticsBodyWithSqon:
      properties:
        field:
          type: string
        sqon:
          $ref: '#/components/schemas/Sqon'
      type: object
...
    VepImpact:
      description: TODO
      enum:
      - MODIFIER
      - LOW
      - MODERATE
      - HIGH
      type: string
      x-enum-varnames:
      - MODIFIER
      - LOW
      - MODERATE
      - HIGH
...
  securitySchemes:
    bearerauth:
      bearerFormat: JWT
      scheme: bearer
      type: http
...
info:
  description: This is the API for Radiant data platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Radiant API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /occurrences/somatic/snv/{case_id}/{seq_id}/{locus_id}/expanded:
    get:
      description: Retrieve ExpandedSomaticSNVOccurrence data for a given locus ID
      operationId: getExpandedSomaticSNVOccurrence
      parameters:
      - description: Case ID
        in: path
        name: case_id
        required: true
        schema:
          type: integer
      - description: Sequence ID
        in: path
        name: seq_id
        required: true
        schema:
          type: integer
      - description: Locus ID
        in: path
        name: locus_id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpandedSomaticSNVOccurrence'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal Server Error
      security:
      - bearerauth: []
      summary: Get a somatic ExpandedSomaticSNVOccurrence
      tags:
      - occurrences
  /occurrences/somatic/snv/{case_id}/{seq_id}/aggregate:
    post:
      description: Aggregate somatic SNV occurrences for a given sequence ID
      operationId: aggregateSomaticSNVOccurrences
      parameters:
      - description: Case ID
        in: path
        name: case_id
        required: true
        schema:
          type: integer
      - description: Sequence ID
        in: path
        name: seq_id
        required: true
        schema:
          type: integer
      - description: Whether to include all possible facet values
        in: query
        name: with_dictionary
        schema:
          default: false
          type: boolean
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AggregationBodyWithSqon'
        description: Aggregation Body
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Aggregation'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal Server Error
      security:
      - bearerauth: []
      summary: Aggregate somatic SNV occurrences
      tags:
      - occurrences
  /occurrences/somatic/snv/{case_id}/{seq_id}/count:
    post:
      description: Counts somatic SNV occurrences for a given sequence ID
      operationId: countSomaticSNVOccurrences
      parameters:
      - description: Case ID
        in: path
        name: case_id
        required: true
        schema:
          type: integer
      - description: Sequence ID
        in: path
        name: seq_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CountBodyWithSqon'
        description: Count Body
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Count'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal Server Error
      security:
      - bearerauth: []
      summary: Count somatic SNV occurrences
      tags:
      - occurrences
  /occurrences/somatic/snv/{case_id}/{seq_id}/list:
    post:
      description: List somatic SNV occurrences for a given sequence ID
      operationId: listSomaticSNVOccurrences
      parameters:
      - description: Case ID
        in: path
        name: case_id
        required: true
        schema:
          type: integer
      - description: Sequence ID
        in: path
        name: seq_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListBodyWithSqon'
        description: List Body
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SomaticSNVOccurrence'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal Server Error
      security:
      - bearerauth: []
      summary: List somatic SNV occurrences
      tags:
      - occurrences
  /occurrences/somatic/snv/{case_id}/{seq_id}/statistics:
    post:
      description: Return statistics about a field for a given sequence ID
      operationId: statisticsSomaticSNVOccurrences
      parameters:
      - description: Case ID
        in: path
        name: case_id
        required: true
        schema:
          type: integer
      - description: Sequence ID
        in: path
        name: seq_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatisticsBodyWithSqon'
        description: Statistics Body
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Statistics'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal Server Error
      security:
      - bearerauth: []
      summary: Statistics of somatic SNV occurrences
      tags:
      - occurrences
...
servers:
- url: /